+++ src/hooks/useDropDown.ts
@@ -1,0 +1,32 @@
+import { useEffect, useRef, useState } from "react";
+
+function useDropDown() {
+  const [isOpen, setIsOpen] = useState(false);
+  const dropdownRef = useRef<HTMLDivElement>(null);
+
+  const toggleDropdown = () => setIsOpen(!isOpen);
+  const closeDropdown = () => setIsOpen(false);
+
+  useEffect(() => {
+    const handleClickOutside = (event: MouseEvent) => {
+      if (
+        dropdownRef.current &&
+        !dropdownRef.current.contains(event.target as Node)
+      ) {
+        closeDropdown();
+      }
+    };
+
+    document.addEventListener("click", handleClickOutside);
+    return () => document.removeEventListener("click", handleClickOutside);
+  }, []);
+
+  return {
+    isOpen,
+    dropdownRef,
+    toggleDropdown,
+    closeDropdown,
+  };
+}
+
+export default useDropDown;
